requirements.txt
asgiref==3.5.2
async-generator==1.10
attrs==22.1.0
certifi==2022.9.24
cffi==1.15.1
Django==4.1.1
django-crispy-forms==1.14.0
django-filter==22.1
exceptiongroup==1.0.4
h11==0.14.0
idna==3.4
outcome==1.2.0
pycparser==2.21
PySocks==1.7.1
selenium==4.7.2
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.3
trio==0.22.0
trio-websocket==0.9.2
tzdata==2022.7
urllib3==1.26.13
whitenoise==6.6.0
wsproto==1.2.0
